+++ CHANGELOG.md
@@ -4,4 +4,6 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ## 1.0.0 (2021-12-15)
 
+* Add first working implementation to report updates based on a point in time.
+